cf: add workers entry packages/api/src/worker.ts
- Extract Hono app construction into app.ts (runtime-agnostic).
- Bun entry index.ts now imports app, mounts the in-memory rate-limit log,
  /ready probe, runMigrations(), Bun.serve and SIGINT/SIGTERM handlers.
- worker.ts is the Cloudflare Workers fetch shim. No setInterval, no
  runMigrations, no signal handlers, no top-level network awaits.
- Hydrates Worker env bindings into process.env so request-time secret
  reads (JWT, RPC URL, etc.) keep working under nodejs_compat.
